Incapsulated messages functional into separate component
